move color naming into its own file
